requirements.txt
Flask
sqlalchemy
pymysql
python-dotenv
gunicorn
